+++ visualization/predictions/prediction_options.py
@@ -14,8 +14,8 @@ class PredictionOptions(QWidget):
         centerPoint = QtWidgets.QDesktopWidget().availableGeometry().center()
         self.width = int(parent.width / 2)
         self.height = int(parent.height / 2.5)
-        self.left = centerPoint.x() - self.width / 2
-        self.top = centerPoint.y() - self.height / 2
+        self.left = int(centerPoint.x() - self.width / 2)
+        self.top = int(centerPoint.y() - self.height / 2)
         self.title = 'Prediction Options'
         self.data = pi
         self.parent = parent
